App channel revision (#9)
* app channel revision
mostly adaptions that were brought over from the revised payment channel example to attain similarities between these examples

* updated ci

* fix: register app

* fix: wait for app deployment

* deploy: panic on error

* OpenAppChannel: Improve documentation

* OpenAppChannel: Might be cleaner to return pointer

* CI: Move app channels before collateralized channels

* CI: don't fail if container is not running

* moved grid comment

* added comments to channel.go

* Readme: Removed ganache --gasPrice

* ForceSet: Added for dispute case

* fix: TicTacToeApp.sol winner index

* main.go: examplary dispute case

* AppChannel renamed  toTicTacToeChannel

* CI: removed --gasPrice

* Client: payment -> app in comments

* Scenario: Dispute trigger

* Revert "fix: TicTacToeApp.sol winner index"

This reverts commit 77e2aa794ee9d4f3700f7afe14f62b7579727ac8.

* fix: TicTacToeApp.sol winner index requirement

* Scenario: Removed time.sleep
and made dispute case standard

* Payment Channel: OpenChannel now return a pointer
as in the app channel example

* Deployment: Changed gas limit
to be closer to the TicTacToe contract requirement (which is currently around 1034584)

* Minor changes through review

Update app-channel/channel.go
Update app-channel/util.go
Update app-channel/main.go
Update app-channel/client/client.go
Update app-channel/client/channel.go

+++ app-channel/app/data.go
@@ -10,6 +10,10 @@ import (
 )
 
 // TicTacToeAppData is the app data struct.
+// Grid:
+// 0 1 2
+// 3 4 5
+// 6 7 8
 type TicTacToeAppData struct {
 	NextActor uint8
 	Grid      [9]FieldValue
@@ -42,7 +46,7 @@ func (d *TicTacToeAppData) Clone() channel.Data {
 }
 
 func (d *TicTacToeAppData) Set(x, y int, actorIdx channel.Index) {
-	if d.NextActor != uint8safe(actorIdx) {
+	if d.NextActor != uint8safe(uint16(actorIdx)) {
 		panic("invalid actor")
 	}
 	v := makeFieldValueFromPlayerIdx(actorIdx)
